engine: add engine trait and hide kernel from users
