+++ src/engine/hash/src/error.rs
@@ -12,9 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use engula_journal::Error as JournalError;
-use engula_kernel::Error as KernelError;
-use engula_storage::Error as StorageError;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
@@ -24,16 +21,16 @@ pub enum Error {
     #[error("corrupted: {0}")]
     Corrupted(String),
     #[error(transparent)]
-    Kernel(#[from] KernelError),
-    #[error(transparent)]
-    Journal(#[from] JournalError),
-    #[error(transparent)]
-    Storage(#[from] StorageError),
+    Unknown(Box<dyn std::error::Error + Send + 'static>),
 }
 
 impl Error {
     pub fn corrupted<E: ToString>(err: E) -> Self {
         Self::Corrupted(err.to_string())
+    }
+
+    pub fn unknown<E: std::error::Error + Send + 'static>(err: E) -> Self {
+        Self::Unknown(Box::new(err))
     }
 }
