+++ src/engine/hash/src/local/codec.rs
